Added internal links to research, etc
On narrow screens, the front page stacks the items, which we want.
But then it's not necessarily clear that teaching, research, etc.
are further down on the page, especially since the 'more information'
link is above them. So I've added links to those sections to make
their existence clear (I hope).